# src/main/resources/META-INF/resources/api.yml
openapi: 3.0.1
info:
  title: contact-service API
  version: 1.0.0
  description: |
    Contact Service
servers:
  - url: http://localhost:8074
    description: Generated server url
tags:
  - name: contact-service
    description: contact-service
  - name: Beneficiary Account
    description: Beneficiary Account
paths:
  /contacts:
    get:
      tags:
        - contact
      summary: Get the list of contact by user identity
      operationId: getContactList
      parameters:
        - name : userId
          in : query
          description: user id
          required: false
          schema:
            type: string
          example: 123
        - name : searchText
          in : query
          description: search text
          required: false
          schema:
            type: string
          example: abc
      responses:
        503:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        200:
          description: Retrieve data succesfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeneficiaryResponse'
        404:
          description: Data Not Found
          content:
            application/json:
              schema:
                type: string
        400:
          description: Input data is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
      security:
        - Bearer: [ ]
    post:
      tags:
        - contact
      summary: Create a new contact for given user identity
      operationId: createContact
      requestBody:
        description: Contact data model
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeneficiaryRecord'
      responses:
          503:
            description: Internal server error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          201:
            description: Contact data model created succesfully
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/BeneficiaryCreateResponse'
          404:
            description: Data Not Found
            content:
              application/json:
                schema:
                  type: string
          400:
            description: Input data is invalid
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          401:
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: string
      security:
        - Bearer: [ ]
  /beneficiaries/-/mobile-numbers/{mobileNumber}:
    get:
      tags:
        -  Beneficiary Account
      summary: Lookup beneficiary account by mobile number
      description: Lookup beneficiary account by mobile number
      operationId:  lookupBeneficiaryAccountWithMobile
      parameters:
        - name: mobileNumber
          in: path
          description: Mobile Number
          required: true
          schema:
            type: string
          example: 65086508427
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeneficiaryAccountResponse'
        503:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Data Not Found
          content:
            application/json:
              schema:
                type: string
        400:
          description: Input data is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
      security:
        - Bearer: [ ]
  /beneficiaries/-/account-numbers/{accountNumber}:
    get:
      tags:
        - Beneficiary Account
      summary: Lookup beneficiary account by account number
      description: Lookup beneficiary account by account number
      operationId:  lookupBeneficiaryAccount
      parameters:
        - name: accountNumber
          in: path
          description: Account Number
          required: true
          schema:
            type: string
          example: 120086508427
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeneficiaryAccountResponse'
        503:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Data Not Found
          content:
            application/json:
              schema:
                type: string
        400:
          description: Input data is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
      security:
        - Bearer: [ ]
  /beneficiaries:
    get:
      tags:
        - Beneficiary Account
      summary: Lookup beneficiary account
      description: Lookup beneficiary account
      operationId:  lookupBeneficiary
      parameters:
        - name: mobileNumber
          in: query
          description: Mobile Number
          required: false
          schema:
            type: string
          example: 65086508427
        - name: accountNumber
          in: query
          description: Account Number
          required: false
          schema:
            type: string
          example: 120086508427
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        503:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Data Not Found
          content:
            application/json:
              schema:
                type: string
        400:
          description: Input data is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
      security:
        - Bearer: [ ]
#-----
components:
  schemas:
    BeneficiaryCreateResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/BeneficiaryRecord'
    BeneficiaryResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BeneficiaryRecord'
        pagging:
          $ref: '#/components/schemas/PagingInformation'
        status:
          $ref: '#/components/schemas/ResponseStatus'
    ServiceDataResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: string
        pagging:
          $ref: '#/components/schemas/PagingInformation'
        status:
          $ref: '#/components/schemas/ResponseStatus'
    PagingInformation:
      type: object
      properties:
        totalRecords:
          type: number
        pageSize:
          type: integer
        pageNumber:
          type: integer
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ResponseStatus'
    ResponseStatus:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    BeneficiaryAccountResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/BeneficiaryAccount'
    WalletResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/WalletDto'
    BeneficiaryAccount:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
          readOnly: true
        accountNumber:
          type: string
          readOnly: true
        displayName:
          type: string
    BeneficiaryRecord:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        serviceCode:
          type: string
        subServiceCode:
          type: string
        displayName:
          type: string
        paymentReference:
          type: string
        mobileNumber:
          type: string
        accountNumber:
          type: string
        branchCode:
          type: string
        bankCode:
          type: string
        city:
          type: string
        state:
          type: string
        postCode:
          type: string
        address:
          type: string
        verificationStatus:
          type: string
        verificationAt:
          type: string
    BankAccoutDto:
      type: object
      required:
        - accountHolderName
        - accountId
        - accountNumber
        - accountSubType
        - accountType
        - balanceType
        - bankBranchId
        - bankBranchName
        - bankCode
        - bankLogo
        - bankName
        - consentId
        - countryCode
        - creditDebitIndicator
        - currencyCode
        - internalProductCategory
        - productId
    WalletDto:
      type: object
      properties:
        walletId:
          type: string
          example: 11889-99982-10056-51950
          description: Wallet ID of the specific wallet
        userId:
          type: string
          example: '2222056798197'
          description: Id of the owner of the wallet(member Id)
        walletName:
          type: string
          example: my wallet 003
          description: Name of the wallet
        currentBalance:
          type: number
          example: 200.23
          description: Current balance of the wallet. Meaningful only when wallet is a type of virtual wallet
        availableBalance:
          type: number
          example: 100.23
          description: Available balance of the wallet. Meaningful only when wallet is a type of virtual wallet
        minimumBalance:
          type: number
          example: 10.23
          description: Minimum balance that this wallet should hold. Meaningful only when wallet is a type of virtual wallet
        bankAccount:
          example: '10.23'
          description: Minimum balance that this wallet should hold. Meaningful only when wallet is a type of virtual wallet
          $ref: '#/components/schemas/BankAccoutDto'
        currencyCode:
          type: string
          example: GBP
          description: Currency code(in ISO 4217 alphabetic code standards) of the wallet. Meaningful only when wallet is a type of virtual wallet
        status:
          type: string
          example: ACTIVE
          description: Status of the wallet
          enum:
            - ACTIVE
            - INACTIVE
            - BLOCKED
            - EXPIRED
            - TXN_SYNCING
        type:
          type: string
          example: BANK_WALLET
          description: Type of the wallet
          enum:
            - BANK_WALLET
            - VIRTUAL_WALLET
        isDefaultWallet:
          type: boolean
          example: true
          description: Flag for indicating the default wallet among all the other wallets of a specific member
        balanceAsAtDateTime:
          type: string
          format: date-time
          example: '2019-12-16T03:37:17Z'
          description: Indicate the last checked time of the balance
        expiredAt:
          type: string
          format: date-time
          example: '2019-12-16T03:37:17Z'
          description: Wallet expiring date(in ISO 8061 format)in UTC
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer